[Move] Gameobject data containers moved to their own files
