+++ HierarchyDecorator/Scripts/Editor/ComponentItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 192aebadabb1aec469ee02046f2367a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
