+++ HierarchyDecorator/Scripts/Editor/HierarchyItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a93f12baa008f27449b145da87efb41e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
